Battery script now scales the battery capacites
A previous problem was that the two batteries in my thinkpad were
treated equally. A battery with a capacity of ~1800 mAh and one with
~6000 mAh are not equally charged when saying both have 50% battery
power remaining.

This script now fixes this issue (for me) with hopefully any number of
batteries.

It scales the current battery power with the design power.

And I have rewritten this in python3 because of obvious reasons. I don't
want to do this in (ba)sh.